requirements.txt
fastapi==0.115.12
httpx==0.28.1
pydantic>=2.12,<2.13
pytest==8.3.5
pytest-asyncio>=1.3,<1.4
uvicorn[standard]==0.34.0
PyYAML==6.0.2
